refactor(cache): move accesor methods back to root `impl` block
Moves the accessor methods back to the root level `impl` block, which
makes rustdoc only show one impl block for consistency.

Also adds `#[doc(hidden)]` and `#[deprecated]` on `user_ref` as this was
included by mistake in a previous PR.

Part of #937.
