+++ cache/in-memory/src/event/role.rs
@@ -1,5 +1,4 @@
 use crate::{config::ResourceType, InMemoryCache, UpdateCache};
-use std::collections::HashSet;
 use twilight_model::{
     gateway::payload::{RoleCreate, RoleDelete, RoleUpdate},
     guild::Role,
@@ -7,25 +6,6 @@ use twilight_model::{
 };
 
 impl InMemoryCache {
-    /// Gets the set of roles in a guild.
-    ///
-    /// This is a O(m) operation, where m is the amount of roles in the guild.
-    /// This requires the [`GUILDS`] intent.
-    ///
-    /// [`GUILDS`]: ::twilight_model::gateway::Intents::GUILDS
-    pub fn guild_roles(&self, guild_id: GuildId) -> Option<HashSet<RoleId>> {
-        self.0.guild_roles.get(&guild_id).map(|r| r.clone())
-    }
-
-    /// Gets a role by ID.
-    ///
-    /// This is an O(1) operation. This requires the [`GUILDS`] intent.
-    ///
-    /// [`GUILDS`]: ::twilight_model::gateway::Intents::GUILDS
-    pub fn role(&self, role_id: RoleId) -> Option<Role> {
-        self.0.roles.get(&role_id).map(|r| r.data.clone())
-    }
-
     pub(crate) fn cache_roles(&self, guild_id: GuildId, roles: impl IntoIterator<Item = Role>) {
         for role in roles {
             self.cache_role(guild_id, role);
